added possibility to edit products and cattegories
-i also start working on countring how many products are in the one concrete category

+++ upload_category.php
@@ -18,6 +18,7 @@ if ($conn->connect_error) {
 }
 
 ?>
+<div class="container">
 
 
 <a href="upload_products.php">upload product</a>
@@ -72,15 +73,18 @@ if (isset($_POST['submit'])) {
             echo "Category ID: " . $row["category_id"] . "<br>";
             echo "Category Name: " . $row["category_name"] . "<br>";
             echo '<img src="data:image/jpeg;base64,'.base64_encode( $row['category_image'] ).'"/>';
-            
+
             //update category ID and name
             echo "
-            <form action='update_category.php' method='POST'>
+            <form action='upload_category.php' method='POST'>
                 <input type='hidden' name='category_id' value='". $row['category_id'] ."'>
+
                 <label for='new_category_id'>New Category ID:</label>
                 <input type='text' id='new_category_id' name='new_category_id' value='". $row['category_id'] ."'><br>
+
                 <label for='new_category_name'>New Category Name:</label>
-                <input type='text' id='new_category_name' name='new_category_name' value='". $row['category_name'] ."'><br>
+                <input type='text' name='new_category_name' value='". $row['category_name'] ."'><br>
+
                 <input type='submit' name='update' value='Update'>
             </form>";
             
@@ -96,20 +100,31 @@ if (isset($_POST['submit'])) {
     }else {
         echo "No categories found.";
     }
+
+    if (isset($_POST['update'])) {
+        $category_id = $_POST['category_id'];
+        $new_category_name = $_POST['new_category_name'];
+        $new_category_id = $_POST['new_category_id'];
+
+        $query_update = "UPDATE product_category SET category_id='$new_category_id', category_name='$new_category_name' WHERE category_id = $category_id";
+        mysqli_query($conn, $query_update);
+        echo "<meta http-equiv='refresh' content='0'>";
+        exit;
+    }
+
+
     if (isset($_POST['delete'])) {
         $category_id = $_POST['category_id'];
         $query_delete = "DELETE FROM product_category WHERE category_id = $category_id";
         mysqli_query($conn, $query_delete);
         header("Location: upload_category.php");
+        echo "<meta http-equiv='refresh' content='0'>";
         exit;
-    }else{
-        echo "cannot delete";
     }
 
     
-    // Close the database connection
 ?>
-
+</div>
 <?php
     include 'partials/footer.php';
 ?>
